pyproject.toml
[project]
name = "resutil"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "nobkat", email = "katayama@rs.tus.ac.jp" }
]
dependencies = [
    "pyyaml>=6.0.0",
    "rich>=13.7.0",
    "pytest-mock>=3.14.0",
    "gitpython>=3.1.0",
    "boxsdk[jwt]>=3.9.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.1.1",
    "wheel>=0.43.0",
    "twine>=5.0.0",
    "flake8>=7.0.0",
    "black>=24.4.0",
    "pytest-mock>=3.14.0",
    "mypy>=1.10.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/resutil"]

[project.scripts]
resutil = 'resutil.cli:main'

[project.urls]
Homepage = "https://github.com/KatayamaLab/resutil"
Issues = "https://github.com/KatayamaLab/resutil/issues"
